# packages/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Pathfinder API
  description: VEuPathDB Strategy Builder Chatbot API
  version: 1.0.0
paths:
...
  /api/v1/veupathdb/auth/login:
    post:
      tags:
      - veupathdb-auth
      summary: Login With Password
      description: Login via VEuPathDB /login, link internal user, and store auth cookies.
      operationId: login_with_password_api_v1_veupathdb_auth_login_post
      parameters:
      - name: siteId
        in: query
        required: true
        schema:
          type: string
          title: Siteid
      - name: redirectTo
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Redirectto
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/LoginPayload'
              - type: 'null'
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/veupathdb/auth/token:
    post:
      tags:
      - veupathdb-auth
      summary: Accept Token
      description: Accept a VEuPathDB Authorization token and store it as a cookie.
      operationId: accept_token_api_v1_veupathdb_auth_token_post
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/TokenPayload'
              - type: 'null'
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/veupathdb/auth/status:
    get:
      tags:
      - veupathdb-auth
      summary: Auth Status
      description: Return current VEuPathDB auth status.
      operationId: auth_status_api_v1_veupathdb_auth_status_get
      parameters:
      - name: siteId
        in: query
        required: true
        schema:
          type: string
          title: Siteid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    AuthStatusResponse:
      properties:
        signedIn:
          type: boolean
          title: Signedin
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
      type: object
      required:
      - signedIn
      title: AuthStatusResponse
      description: Current auth status response.
    AuthSuccessResponse:
      properties:
        success:
          type: boolean
          title: Success
        authToken:
          anyOf:
          - type: string
          - type: 'null'
          title: Authtoken
      type: object
      required:
      - success
      title: AuthSuccessResponse
      description: Success response, optionally carrying the internal auth token.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginPayload:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginPayload
...
    TokenPayload:
      properties:
        token:
          type: string
          title: Token
      type: object
      required:
      - token
      title: TokenPayload
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    APIKeyCookie:
      type: apiKey
      in: cookie
      name: pathfinder-auth
servers:
- url: http://localhost:8000
  description: Local development
- url: https://api.pathfinder.veupathdb.org
  description: Production